# contracts/openapi-v1.json
{
  "components": {
    "schemas": {
      "DeadLetterResponse": {
        "properties": {
          "dead_letter_id": {
            "title": "Dead Letter Id",
            "type": "string"
          },
          "organization_id": {
            "title": "Organization Id",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "protocol": {
            "title": "Protocol",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "received_at": {
            "format": "date-time",
            "title": "Received At",
            "type": "string"
          },
          "site_id": {
            "title": "Site Id",
            "type": "string"
          },
          "topic": {
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "dead_letter_id",
          "organization_id",
          "site_id",
          "protocol",
          "topic",
          "reason",
          "payload",
          "received_at"
        ],
        "title": "DeadLetterResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthResponse": {
        "properties": {
          "env": {
            "title": "Env",
            "type": "string"
          },
          "persistence_backend": {
            "title": "Persistence Backend",
            "type": "string"
          },
          "service": {
            "title": "Service",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "service",
          "env",
          "persistence_backend"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "IngestItemResponse": {
        "properties": {
          "dead_letter_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dead Letter Id"
          },
          "duplicate_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duplicate Of"
          },
          "outbox_event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outbox Event Id"
          },
          "point_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Point Key"
          },
          "protocol": {
            "title": "Protocol",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "status": {
            "enum": [
              "accepted",
              "duplicate",
              "dead_letter"
            ],
            "title": "Status",
            "type": "string"
          },
          "telemetry_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telemetry Id"
          }
        },
        "required": [
          "status",
          "protocol"
        ],
        "title": "IngestItemResponse",
        "type": "object"
      },
      "IngestSummaryResponse": {
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "integer"
          },
          "dead_letter": {
            "title": "Dead Letter",
            "type": "integer"
          },
          "duplicate": {
            "title": "Duplicate",
            "type": "integer"
          }
        },
        "required": [
          "accepted",
          "duplicate",
          "dead_letter"
        ],
        "title": "IngestSummaryResponse",
        "type": "object"
      },
      "IngestionRequestBody": {
        "properties": {
          "organization_id": {
            "minLength": 1,
            "title": "Organization Id",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "protocol": {
            "minLength": 1,
            "title": "Protocol",
            "type": "string"
          },
          "site_id": {
            "minLength": 1,
            "title": "Site Id",
            "type": "string"
          },
          "topic": {
            "minLength": 1,
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "organization_id",
          "site_id",
          "protocol",
          "topic",
          "payload"
        ],
        "title": "IngestionRequestBody",
        "type": "object"
      },
      "IngestionResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/IngestItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "summary": {
            "$ref": "#/components/schemas/IngestSummaryResponse"
          }
        },
        "required": [
          "summary",
          "items"
        ],
        "title": "IngestionResponse",
        "type": "object"
      },
      "MetricsResponse": {
        "properties": {
          "protocol": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Protocol",
            "type": "object"
          },
          "status": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Status",
            "type": "object"
          }
        },
        "required": [
          "status",
          "protocol"
        ],
        "title": "MetricsResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Wave 3 baseline for protocol ingestion, normalization, deduplication, dead-letter policy, and outbox persistence.",
    "title": "Device Ingestion Service",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/ingestion/dead-letters": {
      "get": {
        "operationId": "list_dead_letters_api_v1_ingestion_dead_letters_get",
        "parameters": [
          {
            "in": "query",
            "name": "organization_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Organization Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "site_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Site Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DeadLetterResponse"
                  },
                  "title": "Response List Dead Letters Api V1 Ingestion Dead Letters Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Dead Letters"
      }
    },
    "/api/v1/ingestion/events": {
      "post": {
        "operationId": "ingest_event_api_v1_ingestion_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestionRequestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Event"
      }
    },
    "/api/v1/ingestion/metrics": {
      "get": {
        "operationId": "metrics_api_v1_ingestion_metrics_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetricsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Metrics"
      }
    },
    "/healthz": {
      "get": {
        "operationId": "health_healthz_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    }
  }
}